server - modularize schema and resolvers replacing moment with date-fns
