+++ server/package.json
@@ -31,12 +31,14 @@
     "axios": "^0.18.0",
     "bcrypt": "^3.0.2",
     "body-parser": "^1.18.3",
+    "date-fns": "^1.30.1",
     "express": "^4.16.4",
     "google-oauth-jwt": "^0.2.0",
     "graphql": "^14.0.2",
     "graphql-tools": "^4.0.3",
     "helmet": "^3.15.0",
     "jsonwebtoken": "^8.4.0",
+    "lodash": "^4.17.15",
     "moment": "^2.23.0",
     "mongoose": "^5.3.14",
     "notifications-node-client": "^4.5.2"
