+++ server/src/resolvers/auth.js
@@ -1,0 +1,1 @@
+const authResolvers = {};
