+++ server/src/resolvers/index.js
@@ -1,0 +1,7 @@
+import { merge } from 'lodash';
+
+import Calendar from './calendar';
+
+const resolvers = Calendar;
+
+export default resolvers;
