+++ server/src/schemas/index.js
@@ -1,0 +1,7 @@
+import { merge } from 'lodash';
+
+import Calendar from './calendar';
+
+const typeDefs = Calendar;
+
+export default typeDefs;
